+++ test/Sentry.Tests/SerializationTests.cs
@@ -1,3 +1,5 @@
+using System.Text.Json;
+using System.Text.Json.Serialization;
 using Sentry.Testing;
 
 namespace Sentry.Tests;
@@ -33,5 +35,29 @@ public class SerializationTests
         yield return new object[] {"nested nullable IntPtr", new {Value = (IntPtr?)3}};
         yield return new object[] {"nested UIntPtr", new {Value = (UIntPtr)3}};
         yield return new object[] {"nested nullable UIntPtr", new {Value = (IntPtr?)3}};
+
+        JsonExtensions.ResetSerializerOptions();
+        new SentryOptions().AddJsonConverter(new CustomObjectConverter());
+        yield return new object[] {"custom object with value", new CustomObject("test")};
+        yield return new object[] {"custom object with null", new CustomObject(null)};
+    }
+
+    public class CustomObject
+    {
+        public CustomObject(string value)
+        {
+            Value = value;
+        }
+
+        internal string Value { get; }
+    }
+
+    public class CustomObjectConverter : JsonConverter<CustomObject>
+    {
+        public override CustomObject Read(ref Utf8JsonReader reader, Type typeToConvert, JsonSerializerOptions options)
+            => new(reader.GetString());
+
+        public override void Write(Utf8JsonWriter writer, CustomObject value, JsonSerializerOptions options)
+            => writer.WriteStringValue(value.Value);
     }
 }